[infiniband] ipoib: Handle QP in SQE state
Message-id: <1433997258-5667-4-git-send-email-honli@redhat.com>
Patchwork-id: 117586
O-Subject: [RHEL-7.2 PATCH 3/6] IB/ipoib: Handle QP in SQE state
Bugzilla: 1230203
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: Erez Shitrit <erezsh@mellanox.com>

As the result of a completion error the QP can moved to SQE state by
the hardware. Since it's not the Error state, there are no flushes
and hence the driver doesn't know about that.

The fix creates a task that after completion with error which is not a
flush tracks the QP state and if it is in SQE state moves it back to RTS.

Signed-off-by: Erez Shitrit <erezsh@mellanox.com>
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit 2c0107309550b332f83f8be366521557f4b3fd5a)
Signed-off-by: Honggang Li <honli@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>