[redhat] check-kabi: stop the build on ABI inconsistencies
Message-id: <522b6f8f6103181433d016ffef11b89e7ae14b9b.1434056672.git.aquini@redhat.com>
Patchwork-id: 117904
O-Subject: [RHEL7 PATCH 1/2] redhat: check-kabi: stop the build on ABI inconsistencies
Bugzilla:
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Frantisek Hrbata <fhrbata@redhat.com>

This patch changes RHEL-7 check-kabi tool in order to avoid the the same glitch
that allowed the inconsistency reported at (RHEL-6) Bug 1054904 to go unnoticed
for so long time. We want to make sure the build will stop for any case of kABI
warnings, such as it does for for the kABI failures.

Signed-off-by: Rafael Aquini <aquini@redhat.com>