[perf] x86: Add 'index' param to get_event_constraint() callback
Message-id: <1434142156-27821-5-git-send-email-jolsa@redhat.com>
Patchwork-id: 118221
O-Subject: [PATCH RHEL7.2 BZ1210494 04/16] perf/x86: Add 'index' param to get_event_constraint() callback
Bugzilla: 1210494
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jerome Marchand <jmarchan@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1210494
https://bugzilla.redhat.com/show_bug.cgi?id=1210494

upstream
========
commit 79cba822443a168c8f7f5b853d9c7225a6d5415e
Author: Stephane Eranian <eranian@google.com>
Date: Mon Nov 17 20:06:56 2014 +0100

description
===========
This patch adds an index parameter to the get_event_constraint()
x86_pmu callback. It is expected to represent the index of the
event in the cpuc->event_list[] array. When the callback is used
for fake_cpuc (evnet validation), then the index must be -1. The
motivation for passing the index is to use it to index into another
cpuc array.

Signed-off-by: Rafael Aquini <aquini@redhat.com>