[perf] x86/intel: Make the HT bug workaround conditional on HT enabled
Message-id: <1434142156-27821-12-git-send-email-jolsa@redhat.com>
Patchwork-id: 118228
O-Subject: [PATCH RHEL7.2 BZ1210494 11/16] perf/x86/intel: Make the HT bug workaround conditional on HT enabled
Bugzilla: 1210494
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jerome Marchand <jmarchan@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1210494
https://bugzilla.redhat.com/show_bug.cgi?id=1210494

upstream
========
commit b37609c30e41264c4df4acff78abfc894499a49b
Author: Stephane Eranian <eranian@google.com>
Date: Mon Nov 17 20:07:04 2014 +0100

description
===========
This patch disables the PMU HT bug when Hyperthreading (HT)
is disabled. We cannot do this test immediately when perf_events
is initialized. We need to wait until the topology information
is setup properly. As such, we register a later initcall, check
the topology and potentially disable the workaround. To do this,
we need to ensure there is no user of the PMU. At this point of
the boot, the only user is the NMI watchdog, thus we disable
it during the switch and re-enable it right after.

Having the workaround disabled when it is not needed provides
some benefits by limiting the overhead is time and space.
The workaround still ensures correct scheduling of the corrupting
memory events (0xd0, 0xd1, 0xd2) when HT is off. Those events
can only be measured on counters 0-3. Something else the current
kernel did not handle correctly.

Signed-off-by: Rafael Aquini <aquini@redhat.com>