[vhost] cross-endian support for legacy devices
Message-id: <1434365355-28354-8-git-send-email-thuth@redhat.com>
Patchwork-id: 118407
O-Subject: [RHEL7.2 kernel PATCH 7/9] vhost: cross-endian support for legacy devices
Bugzilla: 1213658
RH-Acked-by: David Gibson <dgibson@redhat.com>
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>
RH-Acked-by: Jason Wang <jasowang@redhat.com>

From: Greg Kurz <gkurz@linux.vnet.ibm.com>

This patch brings cross-endian support to vhost when used to implement
legacy virtio devices. Since it is a relatively rare situation, the
feature availability is controlled by a kernel config option (not set
by default).

The vq->is_le boolean field is added to cache the endianness to be
used for ring accesses. It defaults to native endian, as expected
by legacy virtio devices. When the ring gets active, we force little
endian if the device is modern. When the ring is deactivated, we
revert to the native endian default.

If cross-endian was compiled in, a vq->user_be boolean field is added
so that userspace may request a specific endianness. This field is
used to override the default when activating the ring of a legacy
device. It has no effect on modern devices.

Signed-off-by: Greg Kurz <gkurz@linux.vnet.ibm.com>
Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
Reviewed-by: Cornelia Huck <cornelia.huck@de.ibm.com>
Reviewed-by: David Gibson <david@gibson.dropbear.id.au>

Bugzilla: 1213658
(cherry picked from commit 2751c9882b947292fcfb084c4f604e01724af804)
Signed-off-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
	drivers/vhost/Kconfig
	drivers/vhost/vhost.c
	drivers/vhost/vhost.h
	Some trivial conflicts due to different contexts of the hunks