[netdrv] i40e: Update VEB's enabled_tc after reconfiguration
Message-id: <1434375794-26787-36-git-send-email-sassmann@redhat.com>
Patchwork-id: 118459
O-Subject: [RHEL7.2 PATCH 035/244] i40e: Update VEB's enabled_tc after reconfiguration
Bugzilla: 1173788
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: Neerav Parikh <neerav.parikh@intel.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1173788
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=9360129

When the port TC configuration changes as a result of DCBx the driver
modifies the enabled TCs for the VEBs it manages. But, in the process
it did not update the enabled_tc value that it caches on a per VEB basis.

So, when the next reconfiguration event occurs where the number of TC
value is same as the value cached in enabled_tc for a given VEB; driver
does not modify it's TC configuration by calling appropriate AQ command
believing it is running with the same configuration as requested.
Now, as the VEB is not actually enabled for the TCs that are there any
TC configuration command for VSI attached to that VEB with TCs that are
not enabled for the VEB fails.

This patch fixes this issue.

Change-ID: Ife5694469b05494228e0d850429ea1734738cf29
Signed-off-by: Neerav Parikh <neerav.parikh@intel.com>
Tested-By: Jack Morgan <jack.morgan@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Signed-off-by: Stefan Assmann <sassmann@redhat.com>
(cherry picked from commit 23cd1f095adf110d118ef972914c714176cd48d0)
Signed-off-by: Rafael Aquini <aquini@redhat.com>