[unisys] add visorbus
Message-id: <1434572862-32439-1-git-send-email-earfvids@redhat.com>
Patchwork-id: 119484
O-Subject: [RHEL7.2 PATCH BZ 1228343 v2] unisys: add visorbus
Bugzilla: 1228343
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Peter Martuccelli <peterm@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1228343
Upstream Status: staging-next
Build Info: https://brewweb.devel.redhat.com/taskinfo?taskID=9374521
Tested: On a Unisys s-Par system by ensuring that visornic would
        fuction appropriately with visorbus infrastructure by
        performing networking operation and checking there result.
        Output log is in the BZ ticket.

backport staging-next commit
56df900cb44d18c3ffaf16e09b83aaf37d912cc5

This patch add the visorbus which is the core module for the Unisys
drivers. This patch also creates the unisys entry in staging to
replicate was is currently in staging-next as of commit message
commit 56df900cb44d18c3ffaf16e09b83aaf37d912cc5.

This will be the first of a series of drivers to add. All the
functionality needed for the Unisys s-Par machine depends on
the presence of visorbus. The additional drivers visornic, visorhba,
and visorhid will require the infrastructure of visorbus for them to
function.

[v2]: My testing description is outdated. Fixed by updating my
      test description. Also added correct config settings.

Cc: Erik Arfvidson <earfvids@redhat.com>
Cc: Seth Jennings <sjenning@redhat.com>
Cc: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>