[fs] gfs2: Don't add all glocks to the lru
Message-id: <358174036.20215429.1434658693920.JavaMail.zimbra@redhat.com>
Patchwork-id: 120038
O-Subject: [RHEL7.2 PATCH] [bz#1232841] GFS2: Don't add all glocks to the lru
Bugzilla: 1232841
RH-Acked-by: Abhijith Das <adas@redhat.com>

Hi,

This patch is for rhbz#1232841:
"GFS2: Performance regression in specsfs related to the glock shrinker"

The upstream patch is here:
https://git.kernel.org/cgit/linux/kernel/git/gfs2/linux-gfs2.git/commit/fs/gfs2?h=for-next&id=e7ccaf5fe1590667b3fa2f8df5c5ec9ba0dc5b85

It was tested on systems dellfsvr (server) and fscli01 (client) using
specsfs on RHEL6. The RHEL7 version is still being tested.

Here is a kernel build with this patch:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=9389005

Patch description:

The glocks used for resource groups often come and go hundreds of
thousands of times per second. Adding them to the lru list just
adds unnecessary contention for the lru_lock spin_lock, especially
considering we're almost certainly going to re-use the glock and
take it back off the lru microseconds later. We never want the
glock shrinker to cull them anyway. This patch adds a new bit in
the glops that determines which glock types get put onto the lru
list and which ones don't.

Regards,

Bob Peterson
Red Hat File Systems

Resolves: rhbz#1232841
Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>