[x86] perf: Fix bug in unused code
Message-id: <1435009078-1938-7-git-send-email-ddugger@redhat.com>
Patchwork-id: 120524
O-Subject: [RHEL7.2 PATCH 6/6 V3 BZ1036948] perf/x86/intel: Fix bug in unused code
Bugzilla: 1036948
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Don Dugger <n0ano@n0ano.com>

This file was picked up from a version of the kernel that assumed the
cqm data was stored in the per_cpu structure (even though that per_cpu
structure wasn't modified with that data - it's a good thing this file
was never compiled for the RHEL kernel).

Now that this file is being used the references to the cqm data have to
point to the rh_per_cpu data structure.

Upstream status: Not needed as upstream modifes the per_cpu structure

Signed-off-by: Don Dugger <donald.d.dugger@intel.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>