[powerpc] PowerNV kernel is not able to manage 16G pages
Message-id: <1435664842-12137-1-git-send-email-lvivier@redhat.com>
Patchwork-id: 121040
O-Subject: [RHEL7.2 kernel PATCH v3] powerpc: PowerNV kernel is not able to manage 16G pages
Bugzilla: 1212273
RH-Acked-by: David Gibson <dgibson@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>

If a user wants to use a 16GB huge page size, which is valid for the
platform (the value is given by the device tree) but not managed by
the kernel, the kernel accepts the value but it is not able to
allocate these pages. When the user tries to reserve some pages
from the kernel command line parameter (hugepage=X), the value is
ignored. But when he sets this number using /proc/sys/vm/nr_hugepages,
he has an error (Invalid argument) he doesn't understand as
/proc/meminfo gives him 16GB as a valid value.

We can either not advertise 16GB huge page size or write the code to
manage 16 GB huge page size. As to write the code is not an easy task
and it seems no customer is asking for that, this patch removes the
MMU_PAGE_16G from the known page size list. When the OF device tree is
scanned, this size is ignored and not added in the mmu_psize_defs table.

This patch is downstream only because we think upstream would prefer to
have the piece of code to manage it rather than a piece of code to
avoid the error.

Signed-off-by: Laurent Vivier <lvivier@redhat.com>

BZ:   https://bugzilla.redhat.com/show_bug.cgi?id=1212273
BREW: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=9439906
Signed-off-by: Rafael Aquini <aquini@redhat.com>