[infiniband] mlx4: Always use the correct port for mirrored multicast attachments
Message-id: <246e7b074ab111d4860871cd9398ec8549bb9711.1435751116.git.avadai@redhat.com>
Patchwork-id: 121206
O-Subject: [Patch RHEL7.2 V2 141/317] IB/mlx4: Always use the correct port for mirrored multicast attachments
Bugzilla: 1164527 1164530 1164531 1164536 1164537
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Honggang Li <honli@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: Moni Shoua <monis@mellanox.com>

When attaching a QP to a multicast address in bonded mode, there was an
assumption that the port of the QP must be #1. This assumption isn't the
case under the flow which enables maximal usage of the physical ports.

Fix it by always checking the port of the original flow and create the
mirrored flow on the other port.

Fixes: c6215745b66a ('IB/mlx4: Load balance ports in port aggregation mode')
Signed-off-by: Moni Shoua <monis@mellanox.com>
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 824c25c1abe70a527646056f6911d181facde9cc)

Signed-off-by: Amir Vadai <avadai@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>