[infiniband] mlx4: Request alias GUID on demand
Message-id: <7c5e4dbcf761268f28e849683c0da2e3c8c4cf00.1435751116.git.avadai@redhat.com>
Patchwork-id: 121273
O-Subject: [Patch RHEL7.2 V2 208/317] IB/mlx4: Request alias GUID on demand
Bugzilla: 1164527 1164530 1164531 1164536 1164537
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Honggang Li <honli@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: Yishai Hadas <yishaih@mellanox.com>

Request GIDs from the SM on demand, i.e., when a VF actually needs them,
and release them when the GIDs are no longer in use.

In cloud environments, this is useful for GID migrations, in which a
GID is assigned to a VF on the destination HCA, while the VF on the
source HCA is shutdown (but the GID was not administratively released).

Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit ee59fa0d7e9af130bfc1b75524e04c101670bd5e)

Signed-off-by: Amir Vadai <avadai@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>