[kvm] ppc: book3s-hv: Don't wake thread with no vcpu on guest IPI
Message-id: <1435772586-23615-7-git-send-email-lvivier@redhat.com>
Patchwork-id: 121414
O-Subject: [RHEL7.2 kernel PATCH v3 06/13] KVM: PPC: Book3S HV: Don't wake thread with no vcpu on guest IPI
Bugzilla: 1213669
RH-Acked-by: Thomas Huth <thuth@redhat.com>
RH-Acked-by: David Gibson <dgibson@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Paul Mackerras <paulus@samba.org>

When running a multi-threaded guest and vcpu 0 in a virtual core
is not running in the guest (i.e. it is busy elsewhere in the host),
thread 0 of the physical core will switch the MMU to the guest and
then go to nap mode in the code at kvm_do_nap.  If the guest sends
an IPI to thread 0 using the msgsndp instruction, that will wake
up thread 0 and cause all the threads in the guest to exit to the
host unnecessarily.  To avoid the unnecessary exit, this arranges
for the PECEDP bit to be cleared in this situation.  When napping
due to a H_CEDE from the guest, we still set PECEDP so that the
thread will wake up on an IPI sent using msgsndp.

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Alexander Graf <agraf@suse.de>
(cherry picked from commit ccc07772c99befeda7a7a4b1d05a6f3b762518c2)

Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
	arch/powerpc/kvm/book3s_hv_rmhandlers.S
Conflicts with context of missing commit (#if..#endif):
    b6c295d KVM: PPC: Book3S HV: Accumulate timing information for real-mode code
Signed-off-by: Laurent Vivier <lvivier@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1213669