[kvm] ppc: book3s-hv: Use decrementer to wake napping threads
Message-id: <1435772586-23615-8-git-send-email-lvivier@redhat.com>
Patchwork-id: 121415
O-Subject: [RHEL7.2 kernel PATCH v3 07/13] KVM: PPC: Book3S HV: Use decrementer to wake napping threads
Bugzilla: 1213669
RH-Acked-by: Thomas Huth <thuth@redhat.com>
RH-Acked-by: David Gibson <dgibson@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Paul Mackerras <paulus@samba.org>

This arranges for threads that are napping due to their vcpu having
ceded or due to not having a vcpu to wake up at the end of the guest's
timeslice without having to be poked with an IPI.  We do that by
arranging for the decrementer to contain a value no greater than the
number of timebase ticks remaining until the end of the timeslice.
In the case of a thread with no vcpu, this number is in the hypervisor
decrementer already.  In the case of a ceded vcpu, we use the smaller
of the HDEC value and the DEC value.

Using the DEC like this when ceded means we need to save and restore
the guest decrementer value around the nap.

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Alexander Graf <agraf@suse.de>
(cherry picked from commit fd6d53b12410b4b73e3996629350dee3f4a7994f)

Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
	arch/powerpc/kvm/book3s_hv_rmhandlers.S
Conflicts with context of missing commit:
    b6c295d KVM: PPC: Book3S HV: Accumulate timing information for real-mode code
Signed-off-by: Laurent Vivier <lvivier@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1213669