[kvm] ppc: book3s-hv: Use bitmap of active threads rather than count
Message-id: <1435772586-23615-9-git-send-email-lvivier@redhat.com>
Patchwork-id: 121416
O-Subject: [RHEL7.2 kernel PATCH v3 08/13] KVM: PPC: Book3S HV: Use bitmap of active threads rather than count
Bugzilla: 1213669
RH-Acked-by: Thomas Huth <thuth@redhat.com>
RH-Acked-by: David Gibson <dgibson@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Paul Mackerras <paulus@samba.org>

Currently, the entry_exit_count field in the kvmppc_vcore struct
contains two 8-bit counts, one of the threads that have started entering
the guest, and one of the threads that have started exiting the guest.
This changes it to an entry_exit_map field which contains two bitmaps
of 8 bits each.  The advantage of doing this is that it gives us a
bitmap of which threads need to be signalled when exiting the guest.
That means that we no longer need to use the trick of setting the
HDEC to 0 to pull the other threads out of the guest, which led in
some cases to a spurious HDEC interrupt on the next guest entry.

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Alexander Graf <agraf@suse.de>
(cherry picked from commit 7d6c40da198ac18bd5dd2cd18628d5b4c615d842)
Signed-off-by: Laurent Vivier <lvivier@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1213669
Signed-off-by: Rafael Aquini <aquini@redhat.com>