[redhat] configs: Enable support for PowerPC Nest (NX) compression acceleration
Message-id: <1435789271-16605-23-git-send-email-gduarte@redhat.com>
Patchwork-id: 121447
O-Subject: [RHEL7.2 PATCH BZ 1221925 22/22] config: Enable support for PowerPC Nest (NX) compression acceleration
Bugzilla: 1221925
RH-Acked-by: Herbert Xu <herbert.xu@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

BZ 1221925
Upstream Status: rhel-only

Enable PowerPC Nest (NX) compression acceleration
(CONFIG_CRYPTO_DEV_NX_COMPRESS) and related configs for PowerPC.

Note 1: Unfortunately, we can't build these as module yet without
complicating the enablement of zswap. The zswap driver requires to be
built-in because it references unexported functions (plus, other more
complicated reasons).  In addition, it requires its configuration to be
done only at boot time.  Part of the configuration is setting its
compressor, and since we want to use the NX 842 hardware compressor, the
nx_compress_powernv driver (and assoicated frontend and crypto drivers)
must be either built-in, or they must be contained in the initramfs.
If these are set as modules, then the user would be required not only to
configure zswap, but also to edit the dracut configuration, and rebuild
the initramfs.  So in the interest of keeping zswap configuration as
easy as possible, it's best to have these configured built-in.

IBM is working on upstream patches to allow zswap's parameters to be
changed at runtime, so the NX 842 drivers won't need to be built-in
after those patches are accepted. They probably won't be upstream in
time for RHEL 7.2, though.

Note 2: We were disabling CONFIG_CRYPTO_DEV_NX on ppc64le, though what
we actually want to disable is CONFIG_CRYPTO_DEV_NX_ENCRYPT, which does
not work on ppc64le. Patch 10/22 in this series (crypto: nx - add NX-842
platform frontend driver) disables CONFIG_CRYPTO_DEV_NX_ENCRYPT on
ppc64le, therefore we don't need to disable it explicitely in our config
files anymore.

Signed-off-by: Rafael Aquini <aquini@redhat.com>