[fs] ovl: add mutli-layer infrastructure
Message-id: <20150703155940.14948.55887.stgit@warthog.procyon.org.uk>
Patchwork-id: 121561
O-Subject: [RHEL7 PATCH 04/28] ovl: add mutli-layer infrastructure
Bugzilla: 1178942
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

From: Miklos Szeredi <mszeredi@suse.cz>

Add multiple lower layers to 'struct ovl_fs' and 'struct ovl_entry'.

ovl_entry will have an array of paths, instead of just the dentry.  This
allows a compact array containing just the layers which exist at current
point in the tree (which is expected to be a small number for the majority
of dentries).

The number of layers is not limited by this infrastructure.

Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>
Original-commit: dd662667e6d3e55b42798a6e6e7f37dddc639460
RH-Bugzilla: 1178942
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>