[fs] ovl: Fix opaque regression in ovl_lookup
Message-id: <20150703160136.14948.22995.stgit@warthog.procyon.org.uk>
Patchwork-id: 121575
O-Subject: [RHEL7 PATCH 18/28] ovl: Fix opaque regression in ovl_lookup
Bugzilla: 1178942
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

From: hujianyang <hujianyang@huawei.com>

Current multi-layer support overlayfs has a regression in
.lookup(). If there is a directory in upperdir and a regular
file has same name in lowerdir in a merged directory, lower
file is hidden and upper directory is set to opaque in former
case. But it is changed in present code.

In lowerdir lookup path, if a found inode is not directory,
the type checking of previous inode is missing. This inode
will be copied to the lowerstack of ovl_entry directly.

That will lead to several wrong conditions, for example,
the reading of the directory in upperdir may return an error
like:

   ls: reading directory .: Not a directory

This patch makes the lowerdir lookup path check the opaque
for non-directory file too.

Signed-off-by: hujianyang <hujianyang@huawei.com>
Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>
Original-commit: a425c037f3dd8a56469158ab5f37beb46402d958
RH-Bugzilla: 1178942
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>