[fs] ovl: discard independent cursor in readdir()
Message-id: <20150703160152.14948.85746.stgit@warthog.procyon.org.uk>
Patchwork-id: 121577
O-Subject: [RHEL7 PATCH 20/28] ovl: discard independent cursor in readdir()
Bugzilla: 1178942
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

From: hujianyang <hujianyang@huawei.com>

Since the ovl_dir_cache is stable during a directory reading, the cursor
of struct ovl_dir_file don't need to be an independent entry in the list
of a merged directory.

This patch changes *cursor* to a pointer which points to the entry in the
ovl_dir_cache. After this, we don't need to check *is_cursor* either.

Signed-off-by: hujianyang <hujianyang@huawei.com>
Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>
Original-commit: 4330397e4e8a662f36d101659e2a59ce32e76ff4
RH-Bugzilla: 1178942
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>