[fs] ovl: allow distributed fs as lower layer
Message-id: <20150703160250.14948.70550.stgit@warthog.procyon.org.uk>
Patchwork-id: 121584
O-Subject: [RHEL7 PATCH 27/28] ovl: allow distributed fs as lower layer
Bugzilla: 1178942
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

From: Miklos Szeredi <mszeredi@suse.cz>

Allow filesystems with .d_revalidate as lower layer(s), but not as upper
layer.

For local filesystems the rule was that modifications on the layers
directly while being part of the overlay results in undefined behavior.

This can easily be extended to distributed filesystems: we assume the tree
used as lower layer is static, which means ->d_revalidate() should always
return "1".  If that is not the case, return -ESTALE, don't try to work
around the modification.

Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>
Original-commit: 7c03b5d45b8eebf0111125053d8fe887cc262ba6
RH-Bugzilla: 1178942, 1238365
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>