[char] ipmi_ssif: add arguments that were removed upstream
Message-id: <1436456445-49833-1-git-send-email-tcamuso@redhat.com>
Patchwork-id: 121840
O-Subject: [PATCH RHEL7.2 45/44] ipmi_ssif: add arguments that were removed upstream
Bugzilla: 1229675
RH-Acked-by: Thomas Huth <thuth@redhat.com>

RHEL7 Only

Upstream 99ab32f3 removed the priority argument from the sender function
pointer field in the exposed struct ipmi_smi_handlers and all
implementations of the sender function.

C calling convention assures that this would not be a problem for third
party software that makes the call with the priority argument, because
it's the last argument in the arg list. However, this could cause
confusion for third party users who expect to see that argument in the
(*sender)() field of struct ipmi_smi_handlers.

For the sake of clarity, we leave that arg in place in the (*sender)()
field in the struct and add it to the implementation of the sender func
in the new driver ipmi_ssif.c. The arg is unused, so it's harmless.

Also had to add the sysfs_name arg to the call to ipmi_register_smi() in
new driver ipmi_ssif.c. ipmi_register_smi() is an exported function and
we can't change this interface for RHEL7. Upstream removed this argument
from the function and all its callers. The sysfs_name arg is a string
constant and is unused, so it is also harmless.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>