[redhat] spec: Update dracut dependency to pull in drbg module
Message-id: <088c51dd03766ea534ecc0519d68ba72171c9569.1436457848.git.aquini@redhat.com>
Patchwork-id: 121842
O-Subject: [RHEL7 PATCH] redhat: spec: Update dracut dependency to pull in drbg module
Bugzilla: 1241344
RH-Acked-by: Kurt Stutsman <kstutsma@redhat.com>
RH-Acked-by: Frantisek Hrbata <fhrbata@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Herbert Xu <herbert.xu@redhat.com>

RHEL-7 Bugzilla #1241344
Brew URL: https://brewweb.devel.redhat.com/taskinfo?taskID=9494760

Recently drbg became the default crypto RNG.  Previously it was
only the default in FIPS mode.  This causes boot failures when
used with disk encryption because the drbg module was not included
in the initramfs by dracut.

Of course this was broken even before the change as it means
you couldn't use disk encryption in FIPS mode.  However, with
this change people who were able to boot with an encrypted disk
would now encounter a boot failure.

dracut has since been fixed to include the drbg module. This patch
changes the kernel package to depend on the new dracut so that
for upgrades we will always have the new dracut which guarantees
that the generated initramfs would contain drbg.

Signed-off-by: Rafael Aquini <aquini@redhat.com>