[x86] revert "pci: Refine the way to release PCI IRQ resources"
Message-id: <20150713202232.34395.85011.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 121980
O-Subject: [PATCH RHEL7.2 BZ1238216 6/6] Revert "x86/PCI: Refine the way to release PCI IRQ resources"
Bugzilla: 1231358 1238216
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=1238216

Description:
------------
Commit b4b55cda5874 (Refine the way to release PCI IRQ resources)
introduced a regression in the PCI IRQ resource management by causing
the IRQ resource of a device, established when pci_enabled_device()
is called on a fully disabled device, to be released when the driver
is unbound from the device, regardless of the enable_cnt.

This leads to the situation that an ill-behaved driver can now make a
device unusable to subsequent drivers by an imbalance in their use of
pci_enable/disable_device().  That is a serious problem for secondary
drivers like vfio-pci, which are innocent of the transgressions of
the previous driver.

Since the solution of this problem is not immediate and requires
further discussion, revert commit b4b55cda5874 and the issue it was
supposed to address (a bug related to xen-pciback) will be taken
care of in a different way going forward.

Reported-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>

Upstream Status:
----------------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=9e8ce4b96b781b003e3174fbbc62e1d4388c8b8f

Backport Notes:
Didn't bring arch/x86/pci/intel_mid_pci.c change back since we don't have this in RHEL 7.x.

----------------
Steve Best

Proposed Patch:
---------------

Signed-off-by: Rafael Aquini <aquini@redhat.com>