[netdrv] bna: revert "Update the Driver and Firmware Version"
Message-id: <1436889142-5112-1-git-send-email-ivecera@redhat.com>
Patchwork-id: 122003
O-Subject: [RHEL7.2 PATCH] bna: Revert "bna: Update the Driver and Firmware Version"
Bugzilla: 1240644
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Michal Schmidt <mschmidt@redhat.com>

BZ: #1240644 - bna driver can't load firmware and is unusable

Description:
This patch reverts recent commit that bumps version of required firmware
file. Unfortunately this version is incompatible with firmware required
by bfa driver. Compatible FW of bfa is not upstream yet and is not
planned to include it in RHEL7.2 so we need to revert this bna bump.

Test status:
Tested inhouse.

Upstream commit:
N/A

Signed-off-by: Ivan Vecera <ivecera@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>