[mm] userfaultfd: add VM_UFFD_MISSING and VM_UFFD_WP
Message-id: <1437174639-28447-12-git-send-email-aarcange@redhat.com>
Patchwork-id: 122107
O-Subject: [RHEL7.2 PATCH 11/39] userfaultfd: add VM_UFFD_MISSING and VM_UFFD_WP
Bugzilla: 965657
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

These two flags gets set in vma->vm_flags to tell the VM common code
if the userfaultfd is armed and in which mode (only tracking missing
faults, only tracking wrprotect faults or both). If neither flags is
set it means the userfaultfd is not armed on the vma.

Acked-by: Pavel Emelyanov <xemul@parallels.com>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>