[mm] userfaultfd: call handle_userfault() for userfaultfd_missing() faults
Message-id: <1437174639-28447-13-git-send-email-aarcange@redhat.com>
Patchwork-id: 122106
O-Subject: [RHEL7.2 PATCH 12/39] userfaultfd: call handle_userfault() for userfaultfd_missing() faults
Bugzilla: 965657
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

This is where the page faults must be modified to call
handle_userfault() if userfaultfd_missing() is true (so if the
vma->vm_flags had VM_UFFD_MISSING set).

handle_userfault() then takes care of blocking the page fault and
delivering it to userland.

The fault flags must also be passed as parameter so the "read|write"
kind of fault can be passed to userland.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>