[mm] userfaultfd: cleanup superfluous _irq locking
Message-id: <1437174639-28447-17-git-send-email-aarcange@redhat.com>
Patchwork-id: 122123
O-Subject: [RHEL7.2 PATCH 16/39] userfaultfd: cleanup superfluous _irq locking
Bugzilla: 965657
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

This leftover shouldn't have caused any malfunction because the loop
either schedules or it re-enables irqs immediately and schedule()
doesn't seem to BUG_ON(irqs_disabled()). However lately we've been
using the non blocking model so the schedule isn't really exercised
here. Regardless of the side effects this must be fixed as it's not ok
to enter schedule with irq disabled and it's not beneficial to toggle
irqs in the first place.

Reported-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>